Melhorias, herança e auto increment
