--- projeto-petme-main/projeto-petme.iml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
-    </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="module-library" exported="">
-      <library>
-        <CLASSES>
-          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES>
-          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/build/java" />
-          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/demo/java" />
-          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/generated/java" />
-          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/legacy/java" />
-          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/main/core-api/java" />
-          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/main/core-impl/java" />
-          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/main/protocol-impl/java" />
-          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/main/user-api/java" />
-          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/main/user-impl/java" />
-          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/test/java" />
-        </SOURCES>
-        <jarDirectory url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0" recursive="false" />
-      </library>
-    </orderEntry>
-    <orderEntry type="library" name="mysql-connector-j-8.0.33" level="project" />
-  </component>
-</module>
