--- projeto-petme-main/src/Cachorro.java
@@ -1,8 +1,0 @@
-public class Cachorro extends Animal{
-
-    private final String ESPECIE = "Cachorro";
-
-    public String getEspecie() {
-        return "Cachorro";
-    }
-}
