--- projeto-petme-main/src/Gato.java
@@ -1,8 +1,0 @@
-public class Gato extends Animal{
-
-    private final String ESPECIE = "Gato";
-
-    public String getEspecie() {
-        return "Gato";
-    }
-}
