+++ projeto-petme.iml
@@ -7,25 +7,35 @@
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="module-library" exported="">
+    <orderEntry type="module-library">
       <library>
         <CLASSES>
-          <root url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0" />
+          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0" />
         </CLASSES>
         <JAVADOC />
         <SOURCES>
-          <root url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0/src/build/java" />
-          <root url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0/src/demo/java" />
-          <root url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0/src/generated/java" />
-          <root url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0/src/legacy/java" />
-          <root url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0/src/main/core-api/java" />
-          <root url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0/src/main/core-impl/java" />
-          <root url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0/src/main/protocol-impl/java" />
-          <root url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0/src/main/user-api/java" />
-          <root url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0/src/main/user-impl/java" />
-          <root url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0/src/test/java" />
+          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/build/java" />
+          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/demo/java" />
+          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/generated/java" />
+          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/legacy/java" />
+          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/main/core-api/java" />
+          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/main/core-impl/java" />
+          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/main/protocol-impl/java" />
+          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/main/user-api/java" />
+          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/main/user-impl/java" />
+          <root url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0/src/test/java" />
         </SOURCES>
-        <jarDirectory url="file://$USER_HOME$/Downloads/mysql-connector-j-8.2.0" recursive="false" />
+        <jarDirectory url="file://$MODULE_DIR$/../mysql-connector-j-8.2.0" recursive="false" />
+      </library>
+    </orderEntry>
+    <orderEntry type="library" name="mysql-connector-j-8.0.33" level="project" />
+    <orderEntry type="module-library" exported="">
+      <library>
+        <CLASSES>
+          <root url="jar://$MODULE_DIR$/src/mysql-connector-j-8.0.33.jar!/" />
+        </CLASSES>
+        <JAVADOC />
+        <SOURCES />
       </library>
     </orderEntry>
   </component>
