+++ src/AnimalDAO.java
@@ -5,47 +5,21 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 
 public class AnimalDAO {
-    private Connection connection;
+    private final Connection connection;
 
     public AnimalDAO() {
-        this.connection = new Conexao().GeraConexao();
-    }
-
-    private boolean animalJaExiste(int idAnimal) {
-        String sql = "SELECT COUNT(*) FROM animais WHERE id_animal = ?";
-        try {
-            PreparedStatement stmt = connection.prepareStatement(sql);
-            stmt.setInt(1, idAnimal);
-            ResultSet resultSet = stmt.executeQuery();
-
-            // Verifica se há pelo menos um resultado no conjunto
-            if (resultSet.next()) {
-                int count = resultSet.getInt(1);
-                stmt.close();
-                return count > 0;
-            } else {
-                stmt.close();
-                return false; // Retorna falso se não houver resultados
-            }
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
+        this.connection = Conexao.GeraConexao();
     }
 
     // Create
     public void adiciona(Animal animal) {
-        if (animalJaExiste(animal.getIdAnimal())) {
-            System.out.println("Animal com ID " + animal.getIdAnimal() + " já existe na tabela.");
-            return;
-        }
 
-        String sql = "INSERT INTO animais(id_animal, nome_animal, especie, coloracao) VALUES(?,?,?,?)";
+        String sql = "INSERT INTO animais(nome_animal, especie, coloracao) VALUES(?,?,?)";
         try {
             PreparedStatement stmt = connection.prepareStatement(sql);
-            stmt.setInt(1, animal.getIdAnimal());
-            stmt.setString(2, animal.getNomeAnimal());
-            stmt.setString(3, animal.getEspecie());
-            stmt.setString(4, animal.getColoracao());
+            stmt.setString(1, animal.getNomeAnimal());
+            stmt.setString(2, animal.getEspecie());
+            stmt.setString(3, animal.getColoracao());
             stmt.execute();
             stmt.close();
             System.out.println("Pet cadastrado com sucesso.");
@@ -57,16 +31,22 @@ public class AnimalDAO {
     // Read
     public ArrayList<Animal> listar() {
         ArrayList<Animal> animais = new ArrayList<>();
-        String sql = "SELECT id_animal, nome_animal, especie, coloracao FROM animais";
+        String sql = "SELECT * FROM animais";
         try {
             PreparedStatement stmt = connection.prepareStatement(sql);
             ResultSet resultSet = stmt.executeQuery();
             while (resultSet.next()) {
-                Animal animal = new Animal();
-                animal.setIdAnimal(resultSet.getInt("id_animal"));
+                Animal animal;
+                if ("Cachorro".equalsIgnoreCase(resultSet.getString("especie"))) {
+                    animal = new Cachorro();
+                } else {
+                    animal = new Gato();
+                }
+
                 animal.setNomeAnimal(resultSet.getString("nome_animal"));
-                animal.setEspecie(resultSet.getString("especie"));
                 animal.setColoracao(resultSet.getString("coloracao"));
+                animal.setIdAnimal(resultSet.getInt("id_animal"));
+
                 animais.add(animal);
             }
             stmt.close();
@@ -78,17 +58,24 @@ public class AnimalDAO {
 
     // Read por ID
     public Animal buscarPorId(int idAnimal) {
-        String sql = "SELECT id_animal, nome_animal, especie, coloracao FROM animais WHERE id_animal = ?";
+        String sql = "SELECT * FROM animais WHERE id_animal = ?";
         try {
             PreparedStatement stmt = connection.prepareStatement(sql);
             stmt.setInt(1, idAnimal);
             ResultSet resultSet = stmt.executeQuery();
             if (resultSet.next()) {
-                Animal animal = new Animal();
-                animal.setIdAnimal(resultSet.getInt("id_animal"));
+
+                Animal animal;
+                if ("Cachorro".equalsIgnoreCase(resultSet.getString("especie"))) {
+                    animal = new Cachorro();
+                } else {
+                    animal = new Gato();
+                }
+
                 animal.setNomeAnimal(resultSet.getString("nome_animal"));
-                animal.setEspecie(resultSet.getString("especie"));
                 animal.setColoracao(resultSet.getString("coloracao"));
+                animal.setIdAnimal(resultSet.getInt("id_animal"));
+
                 stmt.close();
                 return animal;
             } else {
@@ -99,6 +86,7 @@ public class AnimalDAO {
             throw new RuntimeException(e);
         }
     }
+
 
     // Update
     public void atualiza(Animal animal) {
