+++ src/ApadrinhamentoDAO.java
@@ -5,60 +5,21 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 
 public class ApadrinhamentoDAO {
-    private Connection connection;
+    private final Connection connection;
 
     public ApadrinhamentoDAO() {
-        this.connection = new Conexao().GeraConexao();
-    }
-
-//    private boolean apadrinhamentoJaExiste(int idApadrinhamento) {
-//        String sql = "SELECT COUNT(*) FROM apadrinhamento WHERE id_apadrinhamento = ?";
-//        try {
-//            PreparedStatement stmt = connection.prepareStatement(sql);
-//            stmt.setInt(1, idApadrinhamento);
-//            ResultSet resultSet = stmt.executeQuery();
-//            int count = resultSet.getInt(1);
-//            stmt.close();
-//            return count > 0;
-//        } catch (SQLException e) {
-//            throw new RuntimeException(e);
-//        }
-//    }
-
-    private boolean apadrinhamentoJaExiste(int idApadrinhamento) {
-        String sql = "SELECT COUNT(*) FROM apadrinhamento WHERE id_apadrinhamento = ?";
-        try {
-            PreparedStatement stmt = connection.prepareStatement(sql);
-            stmt.setInt(1, idApadrinhamento);
-            ResultSet resultSet = stmt.executeQuery();
-
-            // Verifica se há pelo menos um resultado no conjunto
-            if (resultSet.next()) {
-                int count = resultSet.getInt(1);
-                stmt.close();
-                return count > 0;
-            } else {
-                stmt.close();
-                return false; // Retorna falso se não houver resultados
-            }
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
+        this.connection = Conexao.GeraConexao();
     }
 
     // Create
     public void adiciona(Apadrinhamento apadrinhamento) {
-        if (apadrinhamentoJaExiste(apadrinhamento.getIdApadrinhamento())) { System.out.println("Apadrinhamento com ID " + apadrinhamento.getIdApadrinhamento() + " já existe na tabela.");
-            return;
-        }
 
-        String sql = "INSERT INTO apadrinhamento(id_apadrinhamento, quantia_mensal, id_pessoa, id_animal) VALUES(?,?,?,?)";
+        String sql = "INSERT INTO apadrinhamento(quantia_mensal, id_pessoa, id_animal) VALUES(?,?,?)";
         try {
             PreparedStatement stmt = connection.prepareStatement(sql);
-            stmt.setInt(1, apadrinhamento.getIdApadrinhamento());
-            stmt.setDouble(2, apadrinhamento.getQuantiaMensal());
-            stmt.setInt(3, apadrinhamento.pegarIdInstanciaPessoa());
-            stmt.setInt(4, apadrinhamento.pegarIdInstanciaAnimal());
+            stmt.setDouble(1, apadrinhamento.getQuantiaMensal());
+            stmt.setInt(2, apadrinhamento.pegarIdInstanciaPessoa());
+            stmt.setInt(3, apadrinhamento.pegarIdInstanciaAnimal());
             stmt.execute();
             stmt.close();
             System.out.println("Apadrinhamento cadastrado com sucesso.");
