+++ src/Cachorro.java
@@ -1,0 +1,8 @@
+public class Cachorro extends Animal{
+
+    private final String ESPECIE = "Cachorro";
+
+    public String getEspecie() {
+        return "Cachorro";
+    }
+}
