+++ src/Gato.java
@@ -1,0 +1,8 @@
+public class Gato extends Animal{
+
+    private final String ESPECIE = "Gato";
+
+    public String getEspecie() {
+        return "Gato";
+    }
+}
