+++ src/Main.java
@@ -1,104 +1,82 @@
-import com.sun.nio.sctp.PeerAddressChangeNotification;
-
 import java.util.ArrayList;
-import java.util.Scanner;
+import static Coletar.Coletar.*;
 
 public class Main {
+
     public static void main(String[] args) {
-        Scanner sc1 = new Scanner(System.in); // Para strings
-        Scanner sc2 = new Scanner(System.in); // Para números
 
         boolean flagMenu1 = false;
 
-        do {
-            System.out.println("");
-            System.out.println("=========================");
-            System.out.println("Escolha uma das opções:");
-            System.out.println("1. Padrinho");
-            System.out.println("2. Pet");
-            System.out.println("3. Apadrinhamento");
-            System.out.println("4. Sair");
-            System.out.println("=========================");
-            System.out.println("");
-
-            System.out.print("Sua escolha: ");
-            String escolhaMenu1 = sc1.nextLine();
+        while (!flagMenu1) {
+            System.out.println("""
+                                \n=========================
+                                Escolha uma das opções:
+                                1. Padrinho
+                                2. Pet
+                                3. Apadrinhamento
+                                4. Sair
+                                =========================
+                                """);
+            int escolhaMenu1 = coletarInt
+                    ("Sua escolha: ", 1, 4);
 
             switch(escolhaMenu1) {
-                case "1":
+                case 1 -> {
                     boolean flagMenu2 = false;
                     do {
-                        System.out.println("");
-                        System.out.println("=========================");
-                        System.out.println("O que deseja fazer?");
-                        System.out.println("1. Cadastrar padrinho");
-                        System.out.println("2. Listar padrinhos");
-                        System.out.println("3. Pesquisar padrinho");
-                        System.out.println("4. Atualizar cadastro de padrinho");
-                        System.out.println("5. Excluir cadastro de padrinho");
-                        System.out.println("6. Voltar");
-                        System.out.println("=========================");
-                        System.out.println("");
+                        System.out.println("""
+                                \n=========================
+                                O que deseja fazer?
+                                1. Cadastrar padrinho
+                                2. Listar padrinhos
+                                3. Pesquisar padrinho
+                                4. Atualizar cadastro de padrinho
+                                5. Excluir cadastro de padrinho
+                                6. Voltar
+                                =========================
+                                """);
 
-                        System.out.print("Sua escolha: ");
-                        String escolhaMenu2 = sc1.nextLine();
+                        int escolhaMenu2 = coletarInt
+                                ("Sua escolha: ", 1, 6);
 
                         switch (escolhaMenu2) {
-                            case "1":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("CADASTRAR PADRINHO");
-                                System.out.println("=========================");
-                                System.out.println("");
+                            case 1 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        CADASTRAR PADRINHO
+                                        =========================                            
+                                        """);
 
-                                Pessoa p1 = new Pessoa();
-
-                                p1.setIdPessoa(obterNumeroInteiro(sc2, "Digite o ID do padrinho: "));
-                                System.out.print("Digite o nome do padrinho: ");
-                                p1.setNomePessoa(sc1.nextLine());
-                                System.out.print("Digite a cidade do padrinho: ");
-                                p1.setCidade(sc1.nextLine());
-                                System.out.print("Digite o estado do padrinho: ");
-                                p1.setEstado(sc1.nextLine());
-                                System.out.print("Digite o e-mail do padrinho: ");
-                                p1.setEmail(sc1.nextLine());
-                                System.out.print("Digite o Whatsapp do padrinho: ");
-                                p1.setWhatsapp(sc1.nextLine());
+                                Pessoa p1 = new Pessoa(
+                                        coletarString("Digite o nome do padrinho: "),
+                                        coletarString("Digite a cidade do padrinho: "),
+                                        coletarString("Digite o estado do padrinho: ", 2),
+                                        coletarString("Digite o e-mail do padrinho: "),
+                                        coletarString("Digite o Whatsapp do padrinho: ")
+                                );
 
                                 PessoaDAO pd1 = new PessoaDAO();
                                 pd1.adiciona(p1);
+                            }
+                            case 2 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        LISTAR PADRINHOS
+                                        =========================                            
+                                        """);
 
-                                flagMenu2 = false;
-                                break;
-                            case "2":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("LISTAR PADRINHOS");
-                                System.out.println("=========================");
-                                System.out.println("");
+                                Pessoa.listarPessoas();
 
-                                PessoaDAO pd2 = new PessoaDAO();
+                            }
+                            case 3 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        PESQUISAR PADRINHO
+                                        =========================                              
+                                        """);
 
-                                ArrayList<Pessoa> pessoas = pd2.listar();
-                                for (Pessoa pessoa : pessoas) {
-                                    System.out.println("ID: " + pessoa.getIdPessoa() +
-                                            ", Nome: " + pessoa.getNomePessoa() +
-                                            ", Cidade: " + pessoa.getCidade() +
-                                            ", Estado: " + pessoa.getEstado() +
-                                            ", E-mail: " + pessoa.getEmail() +
-                                            ", Whatsapp: " + pessoa.getWhatsapp());
-                                }
-
-                                flagMenu2 = false;
-                                break;
-                            case "3":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("PESQUISAR PADRINHO");
-                                System.out.println("=========================");
-                                System.out.println("");
-
-                                int idBuscaPessoa = obterNumeroInteiro(sc2, "Digite o ID do padrinho: ");
+                                int idBuscaPessoa = coletarInt
+                                        ("Digite o ID do padrinho: ");
 
                                 PessoaDAO pd3 = new PessoaDAO();
                                 Pessoa pessoaEncontrada = pd3.buscarPorId(idBuscaPessoa);
@@ -114,123 +92,118 @@ public class Main {
                                     System.out.println("Padrinho não encontrada.");
                                 }
 
-                                flagMenu2 = false;
-                                break;
-                            case "4":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("ATUALIZAR CADASTRO DE PADRINHO");
-                                System.out.println("=========================");
-                                System.out.println("");
+                            }
+                            case 4 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        ATUALIZAR CADASTRO DE PADRINHO
+                                        =========================                             
+                                        """);
 
-                                int idAtualizaPessoa = obterNumeroInteiro(sc2, "Digite o ID do padrinho a ser atualizado: ");
+                                int idAtualizaPessoa = coletarInt
+                                        ("Digite o ID do padrinho a ser atualizdo: ");
 
                                 PessoaDAO pd4 = new PessoaDAO();
 
                                 Pessoa pessoaAtt = pd4.buscarPorId(idAtualizaPessoa);
                                 if (pessoaAtt != null) {
-                                    System.out.print("Digite o novo nome: ");
-                                    pessoaAtt.setNomePessoa(sc1.nextLine());
+                                    pessoaAtt.setNomePessoa
+                                            (coletarString("Digite o novo nome: "));
                                     pd4.atualiza(pessoaAtt);
                                 }
 
-                                flagMenu2 = false;
-                                break;
-                            case "5":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("EXCLUIR CADASTRO DE PADRINHO");
-                                System.out.println("=========================");
-                                System.out.println("");
+                            }
+                            case 5 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        EXCLUIR CADASTRO DE PADRINHO
+                                        =========================                             
+                                        """);
 
-                                int idExcluirPessoa = obterNumeroInteiro(sc2, "Digite o ID do padrinho a ser excluído: ");
+                                int idExcluirPessoa = coletarInt
+                                        ("Digite o ID do padrinho a ser excluído: ");
 
                                 PessoaDAO pd5 = new PessoaDAO();
 
                                 pd5.exclui(idExcluirPessoa);
 
-                                flagMenu2 = false;
-                                break;
-                            case "6":
-                                System.out.println("Voltando para o menu anterior");
+                            }
+                            case 6 -> {
+                                System.out.println("Voltando para o menu anterior.");
                                 flagMenu2 = true;
-                                break;
-                            default:
-                                System.out.println("Escolha dentre as opções do menu");
+                            }
                         }
-                    } while (flagMenu2 == false);
+                    } while (!flagMenu2);
 
-                    flagMenu1 = false;
-                    break;
-                case "2":
+                }
+                case 2 -> {
                     boolean flagMenu3 = false;
 
                     do {
-                        System.out.println("");
-                        System.out.println("=========================");
-                        System.out.println("O que deseja fazer?");
-                        System.out.println("1. Cadastrar pet");
-                        System.out.println("2. Listar pet");
-                        System.out.println("3. Pesquisar pet");
-                        System.out.println("4. Atualizar cadastro de pet");
-                        System.out.println("5. Excluir cadastro de pet");
-                        System.out.println("6. Voltar");
-                        System.out.println("=========================");
-                        System.out.println("");
+                        System.out.println("""
+                                            \n=========================
+                                            O que deseja fazer?
+                                            1. Cadastrar pet
+                                            2. Listar pet
+                                            3. Pesquisar pet
+                                            4. Atualizar cadastro de pet
+                                            5. Excluir cadastro de pet
+                                            6. Voltar
+                                            =========================
+                                            """);
 
-                        System.out.print("Sua escolha: ");
-                        String escolhaMenu3 = sc1.nextLine();
+                        int escolhaMenu3 = coletarInt
+                                ("Sua escolha: ", 1, 6);
 
                         switch (escolhaMenu3) {
-                            case "1":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("CADASTRAR PET");
-                                System.out.println("=========================");
-                                System.out.println("");
+                            case 1 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        CADASTRAR PET
+                                        =========================                                          
+                                        """);
 
-                                Animal a1 = new Animal();
+                                int especie = coletarInt
+                                        ("""
+                                                1. Cachorro
+                                                2. Gato
 
-                                a1.setIdAnimal(obterNumeroInteiro(sc2, "Digite o ID do pet: "));
-                                System.out.print("Digite o nome do pet: ");
-                                a1.setNomeAnimal(sc1.nextLine());
-                                System.out.print("Digite a espécie do pet: ");
-                                a1.setEspecie(sc1.nextLine());
-                                System.out.print("Digite a coloração do pet: ");
-                                a1.setColoracao(sc1.nextLine());
+                                                Sua resposta: """,1, 2);
+
+                                Animal a1;
+
+                                if (especie == 1) {
+                                    a1 = new Cachorro();
+                                } else {
+                                    a1 = new Gato();
+                                }
+
+                                a1.setNomeAnimal(coletarString("Digite o nome: "));
+                                a1.setColoracao(coletarString("Digite a coloração: "));
 
                                 AnimalDAO ad1 = new AnimalDAO();
                                 ad1.adiciona(a1);
+                            }
+                            case 2 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        LISTAR PETS
+                                        =========================                                                  
+                                        """);
 
-                                flagMenu3 = false;
-                                break;
-                            case "2":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("LISTAR PETS");
-                                System.out.println("=========================");
-                                System.out.println("");
 
-                                AnimalDAO ad2 = new AnimalDAO();
+                                Animal.listarAnimais();
 
-                                ArrayList<Animal> animais = ad2.listar();
-                                for (Animal animal : animais) {
-                                    System.out.println("ID: " + animal.getIdAnimal() +
-                                            ", Nome: " + animal.getNomeAnimal() +
-                                            ", Espécie: " + animal.getEspecie() +
-                                            ", Coloração: " + animal.getColoracao());
-                                }
+                            }
+                            case 3 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        PESQUISAR PET
+                                        =========================                                                  
+                                        """);
 
-                                flagMenu3 = false;
-                                break;
-                            case "3":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("PESQUISAR PET");
-                                System.out.println("=========================");
-                                System.out.println("");
-
-                                int idBuscaAnimal = obterNumeroInteiro(sc2, "Digite o ID do pet: ");
+                                int idBuscaAnimal = coletarInt
+                                        ("Digite o ID do pet: ");
 
                                 AnimalDAO ad3 = new AnimalDAO();
                                 Animal animalEncontrado = ad3.buscarPorId(idBuscaAnimal);
@@ -244,93 +217,91 @@ public class Main {
                                     System.out.println("Pet não encontrado.");
                                 }
 
-                                flagMenu3 = false;
-                                break;
-                            case "4":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("ATUALIZAR CADASTRO DE PET");
-                                System.out.println("=========================");
-                                System.out.println("");
+                            }
+                            case 4 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        ATUALIZAR CADASTRO DE PET
+                                        =========================                                                 
+                                        """);
 
-                                int idAtualizaAnimal = obterNumeroInteiro(sc2, "Digite o ID do pet a ser atualizado: ");
+                                int idAtualizaAnimal = coletarInt
+                                        ("Digite o ID do pet a ser atualizado: ");
 
                                 AnimalDAO ad4 = new AnimalDAO();
 
                                 Animal animalAtt = ad4.buscarPorId(idAtualizaAnimal);
                                 if (animalAtt != null) {
-                                    System.out.print("Digite o novo nome: ");
-                                    animalAtt.setNomeAnimal(sc1.nextLine());
+                                    animalAtt.setNomeAnimal(coletarString
+                                            ("Digite o novo nome: "));
                                     ad4.atualiza(animalAtt);
                                 }
 
-                                flagMenu3 = false;
-                                break;
-                            case "5":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("EXCLUIR CADASTRO DE PET");
-                                System.out.println("=========================");
-                                System.out.println("");
+                            }
+                            case 5 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        EXCLUIR CADASTRO DE PET
+                                        =========================                                                   
+                                        """);
 
-                                int idExcluirPet = obterNumeroInteiro(sc2, "Digite o ID do pet a ser excluído: ");
+                                int idExcluirPet = coletarInt
+                                        ("Digite o ID do pet a ser excluído: ");
 
                                 AnimalDAO ad5 = new AnimalDAO();
 
                                 ad5.exclui(idExcluirPet);
 
-                                flagMenu3 = false;
-                                break;
-                            case "6":
-                                System.out.println("Voltando para o menu anterior");
+                            }
+                            case 6 -> {
+                                System.out.println("Voltando para o menu anterior.");
                                 flagMenu3 = true;
-                                break;
-                            default:
-                                System.out.println("Escolha dentre as opções do menu");
+                            }
                         }
-                    } while (flagMenu3 == false);
+                    } while (!flagMenu3);
 
-                    flagMenu1 = false;
-                    break;
-                case "3":
+                }
+                case 3 -> {
                     boolean flagMenu4 = false;
 
                     do {
-                        System.out.println("");
-                        System.out.println("=========================");
-                        System.out.println("O que deseja fazer?");
-                        System.out.println("1. Cadastrar apadrinhamento");
-                        System.out.println("2. Listar apadrinhamentos");
-                        System.out.println("3. Pesquisar apadrinhamento");
-                        System.out.println("4. Atualizar cadastro de apadrinhamento");
-                        System.out.println("5. Excluir cadastro de apadrinhamento");
-                        System.out.println("6. Voltar");
-                        System.out.println("=========================");
-                        System.out.println("");
+                        System.out.println("""
+                                            \n=========================
+                                            O que deseja fazer?
+                                            1. Cadastrar apadrinhamento
+                                            2. Listar apadrinhamentos
+                                            3. Pesquisar apadrinhamento
+                                            4. Atualizar cadastro de apadrinhamento
+                                            5. Excluir cadastro de apadrinhamento
+                                            6. Voltar
+                                            =========================
+                                            """);
 
-                        System.out.print("Sua escolha: ");
-                        String escolhaMenu4 = sc1.nextLine();
+                        int escolhaMenu4 = coletarInt
+                                ("Sua escolha: ", 1, 6);
 
                         switch (escolhaMenu4) {
-                            case "1":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("CADASTRAR APADRINHAMENTO");
-                                System.out.println("=========================");
-                                System.out.println("");
+                            case 1 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        CADASTRAR APADRINHAMENTO
+                                        =========================                                        
+                                        """);
 
                                 Apadrinhamento ap = new Apadrinhamento();
 
-                                ap.setIdApadrinhamento(obterNumeroInteiro(sc2, "Digite o ID do apadrinhamento: "));
-                                ap.setQuantiaMensal(obterDouble(sc2, "Digite a quantia mensal: "));
+                                ap.setQuantiaMensal(coletarDouble
+                                        ("Digite a quantia mensal: "));
 
                                 PessoaDAO pd = new PessoaDAO();
                                 ArrayList<Pessoa> pessoas = pd.listar();
+                                Pessoa.listarPessoas();
 
                                 boolean flagId = false;
                                 int contador = 0;
                                 do {
-                                    int idPadrinho = obterNumeroInteiro(sc2, "Digite o ID do padrinho: ");
+                                    int idPadrinho = coletarInt
+                                            ("Digite o ID do padrinho: ");
 
                                     for (Pessoa p : pessoas) {
                                         if (p.getIdPessoa() == idPadrinho) {
@@ -343,18 +314,20 @@ public class Main {
 
                                         if (contador == pessoas.size()) {
                                             System.out.println("ID não encontrado. Tente novamente.");
-                                            flagId = false;
                                         }
                                     }
-                                } while (flagId == false);
+                                } while (!flagId);
 
                                 AnimalDAO ad = new AnimalDAO();
                                 ArrayList<Animal> animais = ad.listar();
 
+                                Animal.listarAnimais();
+
                                 flagId = false;
                                 contador = 0;
                                 do {
-                                    int idPet = obterNumeroInteiro(sc2, "Digite o ID do pet: ");
+                                    int idPet = coletarInt
+                                            ("Digite o ID do pet: ");
 
                                     for (Animal a : animais) {
                                         if (a.getIdAnimal() == idPet) {
@@ -367,22 +340,20 @@ public class Main {
 
                                         if (contador == pessoas.size()) {
                                             System.out.println("ID não encontrado. Tente novamente.");
-                                            flagId = false;
                                         }
                                     }
-                                } while (flagId == false);
+                                } while (!flagId);
 
                                 ApadrinhamentoDAO apd = new ApadrinhamentoDAO();
                                 apd.adiciona(ap);
 
-                                flagMenu4 = false;
-                                break;
-                            case "2":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("LISTAR APADRINHAMENTOS");
-                                System.out.println("=========================");
-                                System.out.println("");
+                            }
+                            case 2 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        LISTAR APADRINHAMENTOS
+                                        =========================                                           
+                                        """);
 
                                 ApadrinhamentoDAO apd2 = new ApadrinhamentoDAO();
 
@@ -394,16 +365,16 @@ public class Main {
                                             ", ID do pet: " + apadrinhamento.getIdAnimalApadrinhado());
                                 }
 
-                                flagMenu4 = false;
-                                break;
-                            case "3":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("PESQUISAR APADRINHAMENTO");
-                                System.out.println("=========================");
-                                System.out.println("");
+                            }
+                            case 3 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        PESQUISAR APADRINHAMENTO
+                                        =========================                                          
+                                        """);
 
-                                int idBuscaApadrinhamento = obterNumeroInteiro(sc2, "Digite o ID do apadrinhamento: ");
+                                int idBuscaApadrinhamento = coletarInt
+                                        ("Digite o ID do apadrinhamento: ");
 
                                 ApadrinhamentoDAO apd1 = new ApadrinhamentoDAO();
                                 Apadrinhamento apadrinhamentoEncontrado = apd1.buscarPorId(idBuscaApadrinhamento);
@@ -417,110 +388,57 @@ public class Main {
                                     System.out.println("Apadrinhamento não encontrado.");
                                 }
 
-                                flagMenu4 = false;
-                                break;
-                            case "4":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("ATUALIZAR CADASTRO DE APADRINHAMENTO");
-                                System.out.println("=========================");
-                                System.out.println("");
+                            }
+                            case 4 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        ATUALIZAR CADASTRO DE APADRINHAMENTO
+                                        =========================                                  
+                                        """);
 
-                                int idAtualizaApadrinhamento = obterNumeroInteiro(sc2, "Digite o ID do apadrinhamento a ser atualizado: ");
+                                int idAtualizaApadrinhamento = coletarInt
+                                        ("Digite o ID do apadrinhamento a ser atualizado: ");
 
                                 ApadrinhamentoDAO apd3 = new ApadrinhamentoDAO();
 
                                 Apadrinhamento apadrinhamentoAtt = apd3.buscarPorId(idAtualizaApadrinhamento);
                                 if (apadrinhamentoAtt != null) {
-                                    apadrinhamentoAtt.setQuantiaMensal(obterDouble(sc2, "Digite a nova quantia mensal: "));
+                                    apadrinhamentoAtt.setQuantiaMensal(coletarDouble
+                                            ("Digite a nova quantia: "));
                                     apd3.atualiza(apadrinhamentoAtt);
                                 }
 
-                                flagMenu4 = false;
-                                break;
-                            case "5":
-                                System.out.println("");
-                                System.out.println("=========================");
-                                System.out.println("EXCLUIR CADASTRO DE APADRINHAMENTO");
-                                System.out.println("=========================");
-                                System.out.println("");
+                            }
+                            case 5 -> {
+                                System.out.println("""
+                                        \n=========================
+                                        EXCLUIR CADASTRO DE APADRINHAMENTO
+                                        =========================                                          
+                                        """);
 
-                                int idExcluirApadrinhamento = obterNumeroInteiro(sc2, "Digite o ID do apadrinhamento a ser excluído: ");
+                                int idExcluirApadrinhamento = coletarInt
+                                        ("Digite o ID do apadrinhamento a ser excluído: ");
 
                                 ApadrinhamentoDAO apd4 = new ApadrinhamentoDAO();
 
                                 apd4.exclui(idExcluirApadrinhamento);
 
-                                flagMenu4 = false;
-                                break;
-                            case "6":
-                                System.out.println("Voltando para o menu anterior");
+                            }
+                            case 6 -> {
+                                System.out.println("Voltando para o menu anterior.");
                                 flagMenu4 = true;
-                                break;
-                            default:
-                                System.out.println("Escolha dentre as opções do menu");
+                            }
                         }
-                    } while (flagMenu4 == false);
+                    } while (!flagMenu4);
 
-                    flagMenu1 = false;
-                    break;
-                case "4":
+                }
+                case 4 -> {
                     System.out.println("Até a próxima!");
                     flagMenu1 = true;
-                    break;
-                default:
-                    System.out.println("Escolha dentre as opções do menu");
-            }
-        } while(flagMenu1 == false);
-    }
-
-    // Método para não permitir a entrada de strings e forçar o usuário a digitar ints
-    private static int obterNumeroInteiro(Scanner scanner, String mensagem) {
-        int numero = 0;
-        boolean entradaValida = false;
-
-        while (!entradaValida) {
-            try {
-                System.out.print(mensagem);
-                String input = scanner.nextLine().trim();
-
-                if (input.isEmpty()) {
-                    throw new java.util.InputMismatchException();
                 }
-
-                numero = Integer.parseInt(input);
-                entradaValida = true;
-            } catch (java.util.InputMismatchException | NumberFormatException e) {
-                System.out.println("Por favor, digite um número válido.");
             }
         }
-        return numero;
-    }
 
-    // Método semelhante para forçar doubles
-    private static double obterDouble(Scanner scanner, String mensagem) {
-        double numero = 0;
-        boolean entradaValida = false;
-
-        while (!entradaValida) {
-            try {
-                System.out.print(mensagem);
-                String input = scanner.nextLine().trim();
-
-                if (input.isEmpty()) {
-                    throw new java.util.InputMismatchException();
-                }
-
-                numero = Double.parseDouble(input);
-                entradaValida = true;
-            } catch (java.util.InputMismatchException | NumberFormatException e) {
-                System.out.println("Por favor, digite um número válido.");
-            }
-        }
-        return numero;
-    }
-
-    private static void padrinhoExiste() {
-
+        fecharScanner();
     }
 }
